Shootables
# Description
Added the shootables interface to set the different shootables that can be fired. It may potentially be wrapped on a scriptable object to add extra shoot data and/or use object pooling for the bullets.

The player now also has a bullet origin that orbits around the player demonstrating the point the bullets generate at and the direction of the aim.

A basic bullet was implemented that modifies health on trigger enter.