Contact Enemy
# Description
Added new nodes and conditions to aid with creating the logic for the AI behaviour tree.

Due to a bug in editor where the object stored in the blackboard variables never returning null when null checking, an extra equality check had to be added as recommended by Unity Staff.

Refactored some of the nodes.