Bug Fixing
# Description
Changed the import settings of the textures to reduce blurriness.

When the action finding a path to the target returns an invalid path, now it results in a failure.

The colour animations now play all of them at once instead of in a sequence.

Pathfind returns fail when the path is empty and it attempts to find a walkable node near the unwalkable node requested.